enterprise/ee-ha-raft-cluster/src/main/java/io/github/onograph/cluster/raft/gateway/gwlb/LoadBalancingResourceIntegrator.java
package io.github.onograph.cluster.raft.gateway.gwlb;

import io.github.onograph.cluster.raft.gateway.gwlb.mods.ServerPoliciesPlugin;
import io.github.onograph.cluster.raft.module.dprot.ClusterStructureService;
import io.github.onograph.config.SettingsDeclarationClusterImpl;
import java.util.HashSet;
import java.util.Iterator;
import java.util.Set;
import org.neo4j.configuration.Config;
import org.neo4j.configuration.SettingConstraint;
import org.neo4j.graphdb.config.Configuration;
import org.neo4j.logging.Log;
import org.neo4j.logging.LogProvider;
import org.neo4j.service.Services;


public class LoadBalancingResourceIntegrator {

  private LoadBalancingResourceIntegrator() {
  }


  private static ClusteringBalancerService _fidpunClusteringBalancerService(String _strNc) {

    Set<String> _setAoString = new HashSet();

    Iterable<ClusteringBalancerService> _iterableCoiaClusteringBalancerService = Services.loadAll(
        ClusteringBalancerService.class);

    Iterator _iterator = _iterableCoiaClusteringBalancerService.iterator();

    while (_iterator.hasNext()) {

      ClusteringBalancerService _puiClusteringBalancerService = (ClusteringBalancerService) _iterator.next();
      if (_puiClusteringBalancerService.plgneString().equals(_strNc)) {
        return _puiClusteringBalancerService;
      }

      _setAoString.add(_puiClusteringBalancerService.plgneString());
    }

    throw new IllegalArgumentException("*** Error: 78a49212-57d5-403b-9662-8bfc04759478");
  }


  private static String _gecnrnmString(Config _config) {
    return _config.get(SettingsDeclarationClusterImpl.settingLoaiuString);
  }


  public static SettingConstraint<String> hapgSettingConstraint() {
    return new SettingConstraint<String>() {
      @Override
      public String getDescription() {
        return io.github.onograph.I18N.format(
            "io.github.onograph.cluster.raft.gateway.gwlb.LoadBalancingResourceIntegrator.getDescription");
      }

      @Override
      public void validate(String _strCn, Configuration _configuration) {
        LoadBalancingResourceIntegrator._fidpunClusteringBalancerService(_strCn);
      }
    };
  }


  public static BalancerService loaBalancerService(ClusterStructureService clusterStructureService,
      Config _config, LogProvider _logProvider, ClusterLeader _lsClusterLeader)
      throws Throwable {

    ClusteringBalancerService _puiClusteringBalancerService = _fidpunClusteringBalancerService(
        _gecnrnmString(_config));
    _puiClusteringBalancerService.iniForClClLoCo(clusterStructureService, _lsClusterLeader,
        _logProvider, _config);
    return _config.get(SettingsDeclarationClusterImpl.settingLoaiuBoolean)
        && !_puiClusteringBalancerService.isSfgpi()
        ? new ServerPoliciesPlugin(_puiClusteringBalancerService) : _puiClusteringBalancerService;
  }


  public static void vaitForCoLo(Config _config, Log _log) {

    ClusteringBalancerService _puiClusteringBalancerService = _fidpunClusteringBalancerService(
        _gecnrnmString(_config));
    _puiClusteringBalancerService.vaitForCoLo(_config, _log);
  }
}
